requirements.txt
Flask==2.0.2
gunicorn==20.1.0
importlib-metadata==4.10.0
keras<2.9, >=2.8.0rc0
pandas==1.4.1
Keras-Preprocessing==1.1.2
Markdown==3.3.6
numpy==1.21.5
Pillow==8.4.0
requests==2.26.0
tensorflow-cpu==2.8.0
urllib3==1.26.7
imutils==0.5.4
nltk==3.7